fixed model

+++ profileService/internal/handler/facultyProfile/model.go
@@ -5,12 +5,12 @@ import (
 )
 
 type AddProfileRequest struct {
-	NameEnglish  string `json:"name_eng"`
-	Email        string `json:"email"`
-	Alias        string `json:"alias"`
-	InstituteIDs []int  `json:"institute_ids"`
-	PositionID   int64  `json:"position_id"`
-	Year         int64  `json:"year"`
+	NameEnglish  string  `json:"name_eng"`
+	Email        string  `json:"email"`
+	Alias        string  `json:"alias"`
+	InstituteIDs []int64 `json:"institute_ids"`
+	PositionID   int64   `json:"position_id"`
+	Year         int64   `json:"year"`
 }
 
 type AddProfileResponse struct {
